refactor: extract audio_logic.py from app.py
  Move compute_volume_db and resample_to_16k to dedicated module.
  Signal math separated from app orchestration.
